Add device and precision global switch
- Move cdtype to `dynamiqs._utils`
- Rename `dynamiqs/utils/array_types.py` to `dynamiqs/utils/jax_utils.py`
- Add `device()` and `precision()` utilities
