+++ dynamiqs/__init__.py
@@ -1,7 +1,6 @@
 from importlib.metadata import version
 
 from . import dark
-from ._utils import *  # TODO: remove, dev purpose only
 from .mesolve import mesolve
 from .options import Options
 from .plots import *
