+++ dynamiqs/utils/__init__.py
@@ -1,4 +1,4 @@
-from .array_types import *
+from .jax_utils import *
 from .operators import *
 from .optimal_control import *
 from .random import *
